Changed descartes_trajectory_planning package to descartes_core (issue #12)
