+++ descartes_core/include/descartes_core/trajectory_pt.h
@@ -29,8 +29,8 @@
 #include <Eigen/Geometry>
 #include <eigen_stl_containers/eigen_stl_vector_container.h>
 #include <vector>
-#include "descartes_trajectory_planning/robot_model.h"
-#include "descartes_trajectory_planning/trajectory_pt_transition.h"
+#include "descartes_core/robot_model.h"
+#include "descartes_core/trajectory_pt_transition.h"
 
 
 namespace descartes_core
