+++ descartes_core/include/descartes_core/trajectory_pt_transition.h
