+++ descartes_core/src/joint_trajectory_pt.cpp
@@ -23,7 +23,7 @@
  */
 
 #include <console_bridge/console.h>
-#include "descartes_trajectory_planning/joint_trajectory_pt.h"
+#include "descartes_core/joint_trajectory_pt.h"
 
 #define NOT_IMPLEMENTED_ERR(ret) logError("%s not implemented", __PRETTY_FUNCTION__); return ret;
