+++ descartes_core/src/planning_graph.cpp
@@ -22,7 +22,7 @@
  *      Author: Dan Solomon
  */
 
-#include "descartes_trajectory_planning/planning_graph.h"
+#include "descartes_core/planning_graph.h"
 
 #include <stdio.h>
 #include <iomanip>
