+++ descartes_core/src/trajectory_pt_transition.cpp
@@ -15,15 +15,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+/*
+ * trajectory_pt_transition.cpp
+ *
+ *  Created on: Jun 5, 2014
+ *      Author: Dan Solomon
+ */
 
-#include "descartes_trajectory_planning/cart_trajectory_pt.h"
-#include <gtest/gtest.h>
+#include "descartes_core/trajectory_pt_transition.h"
 
-
-using namespace descartes_core;
-
-
-TEST(CartTrajPt, construction)
+namespace descartes_core
 {
-CartTrajectoryPt def();
-}
+  TrajectoryPtTransition::TrajectoryPtTransition(): method_(Interpolations::DEFAULT)
+  {
+
+  }
+
+  TrajectoryPtTransition::~TrajectoryPtTransition()
+  {
+
+  }
+
+} /* namespace descartes_core */
