+++ descartes_core/test/cart_trajectory_pt.cpp
@@ -15,25 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/*
- * trajectory_pt_transition.cpp
- *
- *  Created on: Jun 5, 2014
- *      Author: Dan Solomon
- */
 
-#include "descartes_trajectory_planning/trajectory_pt_transition.h"
+#include "descartes_core/cart_trajectory_pt.h"
+#include <gtest/gtest.h>
 
-namespace descartes_core
+
+using namespace descartes_core;
+
+
+TEST(CartTrajPt, construction)
 {
-  TrajectoryPtTransition::TrajectoryPtTransition(): method_(Interpolations::DEFAULT)
-  {
-
-  }
-
-  TrajectoryPtTransition::~TrajectoryPtTransition()
-  {
-
-  }
-
-} /* namespace descartes_core */
+CartTrajectoryPt def();
+}
